Added an actual calendar.
Touching any day of the current month in the current year will bring up the weekly bar graph.

Added a limit bar to the bar graph corresponding to the recommended net calories.

Touching the current day in the bar graph will navigate to the calories list.

Touching any day in the bar graph will show a toast of the day and net calories of that day.